pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "sports-briefing"
version = "0.1.0"
description = "Personalized sports control center — daily briefing across NBA, Premier League, Formula 1, and NHL."
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }

dependencies = [
    "typer>=0.12.0",
    "rich>=13.7.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0",
    "nba_api>=1.4.0",
    "pandas>=2.0.0",
    "fastf1>=3.3.0",
]

[project.scripts]
# After `pip install -e .` you can run: sports-briefing run
sports-briefing = "sports_briefing.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["sports_briefing*"]
